Made changes to HoldLeft
There are a couple of inconsistencies with the IREC code. For instance, the object state machine has a "state time" in this case but the old irec code just has a global variable for constant time. I don't know how that will raise issues. I also commented out some private variables in the header file that seemed to be unused--the project built with them commented out so some analysis of why they are there is probably necessary. In addition, I think the pointer paramters need to be compared and discussed in our larger team meeting since they are inconsitent as well. Oh also! There are some commented out variables in the header "MAX_CONTROLLED_DESCENT_TIME" and "ALT_THRESHOLD_CONTROLLED" that I'm not sure of why they are there as well.