Small bug fix so it compiles
I have actually no Idea why but a variable S in AttitudeEKF.cpp was being linked to a defined variable S in the SPI library, which was keeping it from compiling. Changing S to s disconnected the variables and now it compiles, if you understand why tf that worked lmk.